Revert "LPS-99657 Class that implements ConfigurationBeanDeclaration should be in package *.settings.definition"
This reverts commit 6e923a44cc2964d4859c36bf3c1de81263ce6bbe.
